+++ R/BatchtoolsFutureBackend-class.R
@@ -30,18 +30,27 @@
 #' job-script template as variable `resources`. This is based on how
 #' [batchtools::submitJobs()] works, with the exception for specially
 #' reserved names defined by the \pkg{future.batchtools} package;
+#'
 #' * `resources[["modules"]]` is character vector of Linux environment
 #'   modules to be loaded.
+#'
 #' * `resources[["startup"]]` and `resources[["shutdown"]]` are character
 #'   vectors of shell code to be injected to the job script as-is.
+#'
 #' * `resources[["details"]]`, if TRUE, results in the job script outputting
 #'   job details and job summaries at the beginning and at the end.
+#'
 #' * `resources[["rscript"]]` is an optional character vector specifying
 #'   how the 'Rscript' is launched. The `resources[["rscript_args"]]` field
 #'   is an optional character vector specifying the 'Rscript' command-line
 #'   arguments.
+#'
+#' * `resources[["envs"]]`, is an optional names character vector specifying
+#'   environment variables to be set.
+#'
 #' * `resources[["asis"]]` is a character vector that are passed as-is to
 #'   the job script and are injected as job resource declarations.
+#'
 #' * All remaining `resources` named elements are injected as named resource
 #'   specification for the scheduler.
 #'